requirements.txt
Flask
Flask-Admin
Flask-Babel
# Flask-Bootstrap4
Bootstrap-Flask
Flask-Cache
Flask-FlatPages
Flask-Gravatar
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-PyMongo
Flask-Restless
Flask-SQLAlchemy
Flask-Themes
Flask-Uploads
Flask-WTF
python-dotenv
Jinja2
Markdown
PyYAML
SQLAlchemy
WTForms
Werkzeug
argparse
blinker
pymongo
python-dateutil
PyJWT
pytest